fixing ucr issues

+++ custom/icds_reports/migrations/sql_templates/update_tables23.sql
@@ -6,7 +6,7 @@ ALTER TABLE ccs_record_monthly ADD COLUMN anc_weight smallint;
 ALTER TABLE ccs_record_monthly ADD COLUMN anc_blood_pressure smallint;
 ALTER TABLE ccs_record_monthly ADD COLUMN bp_sys smallint;
 ALTER TABLE ccs_record_monthly ADD COLUMN bp_dia smallint;
-ALTER TABLE ccs_record_monthly ADD COLUMN anc_hemogloblin smallint;
+ALTER TABLE ccs_record_monthly ADD COLUMN anc_hemogloblin decimal;
 ALTER TABLE ccs_record_monthly ADD COLUMN bleeding smallint;
 ALTER TABLE ccs_record_monthly ADD COLUMN swelling smallint;
 ALTER TABLE ccs_record_monthly ADD COLUMN blurred_vision smallint;
@@ -21,4 +21,3 @@ ALTER TABLE ccs_record_monthly ADD COLUMN edd date;
 ALTER TABLE ccs_record_monthly ADD COLUMN delivery_nature smallint;
 ALTER TABLE ccs_record_monthly ADD COLUMN is_ebf smallint;
 ALTER TABLE ccs_record_monthly ADD COLUMN breastfed_at_birth smallint;
-
